Use reference semantics for rolls
Instead of resolving rolls when assigning them, we now resolve them when they're executed.
When assigning a roll, we check that it's not self referential, to avoid infinite recursions when resolving the roll.
This enables things like setting a flat number, having rolls refer to that identifier, and then if that number increases in the game it can be recomputed in the roller by just reassigning the identifier